separate function into own file
